+++ tests/effects/t26193.nim
@@ -1,0 +1,8 @@
+discard """
+errormsg: "'wrapper' can have side effects"
+"""
+
+proc cProc() {.importc.}
+
+func wrapper() =
+  cProc()