+++ tests/effects/t26193_legacy.nim
@@ -1,0 +1,9 @@
+discard """
+matrix: "--legacy:importcNoSideEffect"
+"""
+
+proc cProc() {.importc.}
+
+func wrapper() =
+  cProc()
+